vip convert pages file extensions to ts
