+++ src/models/RootState.ts
@@ -1,0 +1,6 @@
+import User from "./User";
+
+export interface RootState {
+  [x: string]: any;
+  user: User;
+}
