--- src/pages/MessagesPage/index.js
@@ -1,5 +1,0 @@
-import Chat from './../../components/Chat/index';
-import Main from './../../components/Layout/Main';
-const MessagesPage = () => <Main component={Chat} />;
-
-export default MessagesPage;
