+++ src/pages/MessagesPage/index.tsx
@@ -1,0 +1,5 @@
+import Chat from '../../components/Chat/index';
+import Main from '../../components/Layout/Main';
+const MessagesPage = () => <Main component={Chat} />;
+
+export default MessagesPage;
