+++ src/router/ProtectedRoute.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { Navigate, Outlet } from 'react-router-dom';
+import { RootState } from '../models/RootState';
+import User from '../models/User';
+
 
 const ProtectedRoute = () => {
-  const isLoggedIn = useSelector((state) => state.authReducer.isLoggedIn);
+  const isLoggedIn = useSelector<RootState, User>((state) => state.authReducer.isLoggedIn);
   return isLoggedIn ? <Outlet /> : <Navigate to="/login" />;
 };
